fixing the main packet creation/decoding and critical bugs in the sensors

+++ ascendfsw/src/BME680Sensor.cpp
@@ -93,8 +93,9 @@ void BME680Sensor::readDataPacket(uint8_t*& packet) {
     return;
   }
 
-  float data[5] = {bme.temperature, bme.pressure / 100.0, bme.humidity,
-                   bme.gas_resistance, bme.readAltitude(SEALEVELPRESSURE_HPA)};
+  float data[5] = {bme.temperature, (float)(bme.pressure / 100.0), bme.humidity,
+                   (float)(bme.gas_resistance),
+                   bme.readAltitude(SEALEVELPRESSURE_HPA)};
 
   for (int i = 0; i < 5; i++) {
     memcpy(packet, &data[i], sizeof(float));
